+++ R/predicates.R
@@ -1,17 +1,19 @@
-#' Test for Quarter Bounds
+#' Test for Period Bounds
 #'
-#' `is_first_of_quarter()` and `is_last_of_quarter()` check whether a
-#' `Date` is the first or respectively the last day of an (arbitrary) quarter.
-#' `is_quarter_bounds()` checks whether two `Date` vectors mark the bounds of
-#' single quarters
+#' `is_first_of_quarter()`, `is_last_of_quarter()`, `is_first_of_year()` and
+#' `is_last_of_year()` check whether a `Date` is the first or respectively the
+#' last day of a quarter/year. `is_quarter_bounds()` and `is_year_bounds` checks
+#' whether two `Date` vectors mark the bounds of (the same) quarters
 #'
 #' @param x,first,last `Date` vectors
 #'
 #' @return a `logical` vector
 #' @export
+#' @rdname period_bounds
 #'
 #' @examples
 #' x <- as.Date(c("2018-01-01", "2018-03-31", "2018-02-14"))
+#' is_first_of_year(x)
 #' is_first_of_quarter(x)
 #' is_last_of_quarter(x)
 #' is_quarter_bounds(x[[1]], x[[2]])
@@ -29,7 +31,7 @@ is_quarter_bounds <- function(
 
 
 #' @export
-#' @rdname is_quarter_bounds
+#' @rdname period_bounds
 is_first_of_quarter <- function(
   x
 ){
@@ -40,7 +42,7 @@ is_first_of_quarter <- function(
 
 
 #' @export
-#' @rdname is_quarter_bounds
+#' @rdname period_bounds
 is_last_of_quarter <- function(
   x
 ){
@@ -49,8 +51,37 @@ is_last_of_quarter <- function(
 
 
 
+#' @export
+#' @rdname period_bounds
+is_year_bounds <- function(
+  first,
+  last
+){
+  is_first_of_year(first) & last_of_year(first) == last
+}
 
-#' @rdname is_quarter_bounds
+
+
+#' @export
+#' @rdname period_bounds
+
+is_first_of_year <- function(x){
+   r <- as.POSIXlt(x, tz = tz(x))
+   r$mon == 0L & r$mday == 1L
+}
+
+
+#' @export
+#' @rdname period_bounds
+is_last_of_year <- function(x){
+   r <- as.POSIXlt(x, tz = tz(x))
+   r$mon == 11L & r$mday == 31L
+}
+
+
+
+#' @export
+#' @rdname period_bounds
 is_Date <- function(
   x
 ){